refactor solver classes and check for cplex install before importing
